make util a package fix double logging problem
