+++ tardis/__init__.py
@@ -1,6 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 import logging, sys
-from tardis.io.util import ColoredLogger
+from tardis.util.colored_logger import ColoredFormatter, formatter_message
 # Affiliated packages may add whatever they like to this file, but
 # should keep this content at the top.
 # ----------------------------------------------------------------------------
@@ -10,12 +10,16 @@ from ._astropy_init import *
 from tardis.base import run_tardis
 from tardis.io.util import yaml_load_config_file as yaml_load
 
+FORMAT = "[$BOLD%(name)-20s$RESET][%(levelname)-18s]  %(message)s ($BOLD%(filename)s$RESET:%(lineno)d)"
+COLOR_FORMAT = formatter_message(FORMAT, True)
+
 logging.captureWarnings(True)
 logger = logging.getLogger('tardis')
-logging.setLoggerClass(ColoredLogger)
 logger.setLevel(logging.INFO)
-console_handler = logging.StreamHandler()
-console_formatter = logging.Formatter('%(name)s - %(levelname)s - %(message)s')
+
+console_handler = logging.StreamHandler(sys.stdout)
+console_formatter = ColoredFormatter(COLOR_FORMAT)
 console_handler.setFormatter(console_formatter)
-logger.addHandler(sys.stdout)
+
+logger.addHandler(console_handler)
 logging.getLogger('py.warnings').addHandler(console_handler)
