+++ tardis/model/base.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 from astropy import constants, units as u
 
-from tardis.util import quantity_linspace
+from tardis.util.base import quantity_linspace
 from tardis.io.model_reader import read_density_file, read_abundances_file, read_uniform_abundances
 from tardis.io.util import HDFWriterMixin
 from tardis.io.decay import IsotopeAbundances
