+++ tardis/plasma/standard_plasmas.py
@@ -5,7 +5,7 @@ import pandas as pd
 import numpy as np
 
 from tardis.io import atomic
-from tardis.util import species_string_to_tuple
+from tardis.util.base import species_string_to_tuple
 from tardis.plasma import BasePlasma
 from tardis.plasma.properties.property_collections import (basic_inputs,
     basic_properties, lte_excitation_properties, lte_ionization_properties,
